BMC#18433: Ensure correct use of QT_TR_NOOP() and qsTr()
Before commit ad78feaf5f00498c41713848185cc354ab3682b6, qsTr() was
being called on the AppUp string twice, resulting in character
corruption in the resulting string.  So, the fix for BMC#17434
resolved the corruption issue but created a regression in the
process.

This commit restores the qsTr() function, ensuring that
translations for "View all applications" load correctly, and wraps
the AppUp string with QT_TR_NOOP(); this ensures that qsTr() is
only called *once* and that no corruption occurs.

Signed-off-by: Patrick McCarty <patrick.mccarty@intel.com>